requirements.txt
line-bot-sdk==1.8.0
flask==1.0.2
gunicorn

numpy==1.14.2
beautifulsoup4==4.9.3